+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -5,8 +5,8 @@ import jakarta.servlet.ServletRegistration;
 import nextstep.mvc.DispatcherServlet;
 import nextstep.mvc.HandlerAdapterRegistry;
 import nextstep.mvc.HandlerMappingRegistry;
-import nextstep.mvc.controller.tobe.AnnotationHandlerAdapter;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.AnnotationHandlerAdapter;
+import nextstep.mvc.controller.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
