+++ app/src/main/java/com/techcourse/ManualHandlerAdapter.java
@@ -5,7 +5,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.controller.Controller;
 
 public class ManualHandlerAdapter implements HandlerAdapter {
     @Override
