+++ mvc/src/main/java/nextstep/mvc/controller/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import nextstep.web.support.RequestMethod;
